Reorganizing a snarl of tests. Couple other tweaks.
- Separate test file for get_layout_template, render_model, render_list and the render_iter stuff
- One extra test for get_layout_template_name(): now 28 tests
- Dropped the unused `dictionary` param for render_model()
- Dropped an unused variable in RenderIterNode
