+++ armstrong/core/arm_layout/tests/__init__.py
@@ -1,2 +1,4 @@
-from .templatetags import *
-from .utils import *
+from .render_model import *
+from .render_list import *
+from .render_iter import *
+from .get_template_method import *
