--- armstrong/core/arm_layout/tests/templatetags/__init__.py
@@ -1,1 +1,0 @@
-from .layout_helpers import *
