+++ armstrong/core/arm_layout/utils.py
@@ -5,15 +5,15 @@ from django.template.loader import render_to_string
 def get_layout_template_name(model, name):
     ret = []
     for a in model.__class__.mro():
-        if not hasattr(a, "_meta"):
-            continue
-        ret.append("layout/%s/%s/%s.html" % (a._meta.app_label,
-            a._meta.object_name.lower(), name))
+        if hasattr(a, "_meta"):
+            ret.append("layout/%s/%s/%s.html" % \
+                (a._meta.app_label, a._meta.object_name.lower(), name))
     return ret
 
 
-def render_model(object, name, dictionary=None, context_instance=None):
-    dictionary = dictionary or {}
-    dictionary["object"] = object
-    return mark_safe(render_to_string(get_layout_template_name(object, name),
-        dictionary=dictionary, context_instance=context_instance))
+def render_model(context, object, name):
+    return mark_safe(render_to_string(
+                get_layout_template_name(object, name),
+                dictionary={'object': object},
+                context_instance=context
+            ))
